feat: 0101_it4

+++ labwork-5/project/src/main/java/task4/Interfaces/IATMmachine.java
@@ -1,8 +1,10 @@
 package task4.Interfaces;
 
+import java.math.BigDecimal;
+
 public interface IATMmachine {
-    public float checkBalance();
-    public void replenishBalance(float money);
-    public float withdrawBalance(float money);
-    public float makeTransaction(int id, float money);
+    public BigDecimal checkBalance();
+    public void replenishBalance(BigDecimal money);
+    public BigDecimal withdrawBalance(BigDecimal money);
+    public BigDecimal makeTransaction(int id, BigDecimal money);
 }
